+++ Pods/Headers/Public/JWTDecode/JWTDecode-umbrella.h
@@ -1,1 +1,0 @@
-../../../Target Support Files/JWTDecode/JWTDecode-umbrella.h