+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Changed
-- Hotels now sorted by region>country or brand/region/country
-- Restaurants now sorted by region>country>location
+- Sort hotels by region>country or brand/region/country
+- Sort Restaurants region>country>location
+- Report actual login status error
+
+### Fixed
+- Google Maps directions work on first launch
 
 ## [Version 1.2.0, Build 649] - 2019-11-21